refactor: ViewOrderDetailResponse 폴더 옮기고 viewOrderDetail 메소드 정리
